增加redis Geo方法

+++ src/main/java/com/zhangwuji/im/api/entity/IMUserGeoData.java
@@ -31,9 +31,9 @@ public class IMUserGeoData implements Serializable {
 
     private String city;
 
-    private Float lng;
+    private Double lng;
 
-    private Float lat;
+    private Double lat;
 
     private String uids;
